first compiling working full migration of placeLiquidityOrders
