+++ src/index.ts
@@ -12,6 +12,7 @@ export * from "./amounts";
 export * from "./balance_reader";
 export * from "./contracts";
 export * from "./encoding";
+export * from "./exchange";
 export * from "./fraction";
 export * from "./onchain_reading";
 export * from "./orderbook";
