+++ test/utilities.ts
@@ -1,14 +1,11 @@
-import BN from "bn.js";
 import Web3 from "web3";
 import { HttpProvider } from "web3-core";
 import {
   EpochTokenLockerInstance,
   BatchExchangeInstance,
 } from "../types/truffle-typings";
-import { TokenInfo } from "../src";
-import { Deposit, Order } from "./resources/examples/model";
 
-const MAXU32 = new BN(2).pow(new BN(32)).sub(new BN(1));
+import { Deposit, Order } from "./resources/examples/model";
 
 const jsonrpc = "2.0";
 const id = 0;
@@ -175,87 +172,4 @@ export async function placeOrders(
     );
   }
   return orderIds;
-}
-
-export async function fetchTokenInfoFromExchange(
-  exchange: BatchExchangeInstance,
-  tokenIds: number[],
-  fallbackSymbolName = "UNKNOWN",
-  fallbackDecimals = -1,
-): Promise<Map<number, TokenInfo>> {
-  const ERC20 = artifacts.require("ERC20Detailed");
-  // Fetching token data from EVM
-  const tokenObjects: Map<number, TokenInfo> = new Map();
-  for (const id of tokenIds) {
-    const tokenAddress = await exchange.tokenIdToAddressMap(id);
-    let tokenInfo;
-    try {
-      const tokenInstance = await ERC20.at(tokenAddress);
-      tokenInfo = {
-        id: id,
-        symbol: await tokenInstance.symbol.call(),
-        decimals: (await tokenInstance.decimals.call()).toNumber(),
-        address: tokenAddress,
-      };
-    } catch (err) {
-      tokenInfo = {
-        id: id,
-        symbol: fallbackSymbolName,
-        decimals: fallbackDecimals,
-        address: tokenAddress,
-      };
-    }
-    tokenObjects.set(id, tokenInfo);
-    // Found Token ${tokenInfo.symbol} at ID ${tokenInfo.id} with ${tokenInfo.decimals} decimals`
-  }
-  return tokenObjects;
-}
-
-export async function sendLiquidityOrders(
-  instance: BatchExchangeInstance,
-  tokenIds: number[],
-  PRICE_FOR_LIQUIDITY_PROVISION: BN,
-  SELL_ORDER_AMOUNT_OWL: BN,
-  OWL_NUMBER_DIGITS = 18,
-): Promise<void> {
-  const minBuyAmount = [];
-  const validTokenIds = [];
-  const tokenInfo = await fetchTokenInfoFromExchange(instance, tokenIds);
-  for (const tokenId of tokenIds) {
-    const numDigits = tokenInfo.get(tokenId)?.decimals;
-    if (numDigits && numDigits != -1) {
-      validTokenIds.push(tokenId);
-      if (numDigits < OWL_NUMBER_DIGITS) {
-        minBuyAmount.push(
-          SELL_ORDER_AMOUNT_OWL.mul(PRICE_FOR_LIQUIDITY_PROVISION).div(
-            new BN(10).pow(new BN(OWL_NUMBER_DIGITS - numDigits)),
-          ),
-        );
-      } else {
-        minBuyAmount.push(
-          SELL_ORDER_AMOUNT_OWL.mul(PRICE_FOR_LIQUIDITY_PROVISION).mul(
-            new BN(10).pow(new BN(numDigits - OWL_NUMBER_DIGITS)),
-          ),
-        );
-      }
-    }
-  }
-  const numOrders = validTokenIds.length;
-  const batchId = (await instance.getCurrentBatchId()).toNumber();
-  if (numOrders == 0) {
-    // No liquidity orders added, as all tokens sufficiently funded, or their decimals could not be determined",
-    return;
-  }
-  await instance.placeValidFromOrders(
-    validTokenIds, // sellToken
-    Array(numOrders).fill(0), // buyToken
-    Array(numOrders).fill(batchId + 2), // validFrom
-    Array(numOrders).fill(MAXU32), // validTo
-    minBuyAmount, // buyAmount
-    Array(numOrders).fill(SELL_ORDER_AMOUNT_OWL), // sellAmount
-  );
-  // "Placed liquidity sell orders for the following tokens",
-  // await Promise.all(
-  //   validTokenIds.map(async (i) => await tokenInfo.get(i)?.address),
-  // );
 }
